Revert "fix: WS - Not here page appears when trying to navigate to Workspaces."
